code clean up

+++ payment-service/internal/model/payment.go
@@ -12,6 +12,6 @@ type Payment struct {
 	ID      int           `json:"payment_id"`
 	OrderID int           `json:"order_id"`
 	Status  PaymentStatus `json:"status"`
-	Amount  int           `json:"amount"`
+	Amount  float64       `json:"amount"`
 	UserID  int           `json:"user_id"`
 }
